fix(engine): extract rlimit init to platform-specific files for windows cross-compilation
