+++ engine/cmd/ion/rlimit_windows.go
@@ -1,0 +1,5 @@
+//go:build windows
+
+package main
+
+// Windows does not use RLIMIT_NOFILE; no file-descriptor limit adjustment needed.
